Enhance AppRegistry and AppAccount functionality with duration management
- Updated the `AppRegistryBase` and `IAppRegistry` to include a `duration` parameter in app registration and management, allowing for expiration handling.
- Introduced a new `_validateDuration` function to ensure access duration adheres to defined limits.
- Enhanced the `AppAccount` contract with a `getAppExpiration` function to retrieve the expiration of installed apps.
- Adjusted related structures and interfaces to accommodate the new duration feature.
- Updated tests to validate the new duration logic, including checks for valid and invalid durations during app registration and creation.

+++ packages/contracts/src/spaces/facets/executor/IExecutor.sol
@@ -37,7 +37,7 @@ struct Group {
     // Whether the group is active.
     bool active;
     // Timepoint at which the group becomes inactive.
-    uint48 expiration;
+    uint64 expiration;
 }
 
 // Structure that stores the details for a scheduled operation. This structure fits into a
@@ -100,7 +100,7 @@ interface IExecutorBase {
     event OperationExecuted(bytes32 indexed operationId, uint32 nonce);
     event OperationCanceled(bytes32 indexed operationId, uint32 nonce);
     event GroupStatusSet(bytes32 indexed groupId, bool active);
-    event GroupExpirationSet(bytes32 indexed groupId, uint48 expiration);
+    event GroupExpirationSet(bytes32 indexed groupId, uint64 expiration);
 }
 
 interface IExecutor is IExecutorBase {
